+++ listenbrainz/webserver/static/js/src/RecentListens.test.tsx
@@ -1111,188 +1111,3 @@ describe("Pagination", () => {
     });
   });
 });
-
-describe("getFeedback", () => {
-  it("calls the API correctly", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...(JSON.parse(
-          JSON.stringify(recentListensPropsOneListen)
-        ) as RecentListensProps)}
-      />
-    );
-
-    const instance = wrapper.instance();
-    const spy = jest.fn().mockImplementation(() => {
-      return Promise.resolve(getFeedbackByMsidResponse);
-    });
-    // eslint-disable-next-line dot-notation
-    instance["APIService"].getFeedbackForUserForRecordings = spy;
-
-    const result = await instance.getFeedback();
-
-    expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy).toHaveBeenCalledWith(
-      "iliekcomputers",
-      "973e5620-829d-46dd-89a8-760d87076287,"
-    );
-    expect(result).toEqual(getFeedbackByMsidResponse.feedback);
-  });
-
-  it("doesn't call the API if there are no listens", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...{
-          ...(JSON.parse(
-            JSON.stringify(recentListensPropsOneListen)
-          ) as RecentListensProps),
-          listens: undefined,
-        }}
-      />
-    );
-
-    const instance = wrapper.instance();
-
-    const res = getFeedbackByMsidResponse.feedback;
-    const spy = jest.fn().mockImplementation(() => {
-      return Promise.resolve(res);
-    });
-    // eslint-disable-next-line dot-notation
-    instance["APIService"].getFeedbackForUserForRecordings = spy;
-
-    const result = await instance.getFeedback();
-
-    expect(spy).toHaveBeenCalledTimes(0);
-  });
-});
-
-describe("loadFeedback", () => {
-  it("updates the recordingFeedbackMap state", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...(JSON.parse(
-          JSON.stringify(recentListensPropsOneListen)
-        ) as RecentListensProps)}
-      />
-    );
-
-    const instance = wrapper.instance();
-
-    instance.getFeedback = jest
-      .fn()
-      .mockImplementationOnce(() =>
-        Promise.resolve(getFeedbackByMsidResponse.feedback)
-      );
-
-    await instance.loadFeedback();
-    expect(wrapper.state("recordingFeedbackMap")).toMatchObject({
-      "973e5620-829d-46dd-89a8-760d87076287": 1,
-    });
-  });
-});
-
-describe("getFeedbackForRecordingMsid", () => {
-  it("returns the feedback after fetching from recordingFeedbackMap state", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...(JSON.parse(
-          JSON.stringify(recentListensPropsOneListen)
-        ) as RecentListensProps)}
-      />
-    );
-
-    const instance = wrapper.instance();
-
-    const recordingFeedbackMap: RecordingFeedbackMap = {
-      "973e5620-829d-46dd-89a8-760d87076287": 1,
-    };
-    wrapper.setState({ recordingFeedbackMap });
-
-    const res = await instance.getFeedbackForRecordingMsid(
-      "973e5620-829d-46dd-89a8-760d87076287"
-    );
-
-    expect(res).toEqual(1);
-  });
-
-  it("returns 0 if the recording is not in recordingFeedbackMap state", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...(JSON.parse(
-          JSON.stringify(recentListensPropsOneListen)
-        ) as RecentListensProps)}
-      />
-    );
-
-    const instance = wrapper.instance();
-
-    const res = await instance.getFeedbackForRecordingMsid(
-      "973e5620-829d-46dd-89a8-760d87076287"
-    );
-
-    expect(res).toEqual(0);
-  });
-});
-
-describe("updateFeedback", () => {
-  it("updates the recordingFeedbackMap state for particular recording", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const wrapper = shallow<RecentListens>(
-      <RecentListens
-        {...(JSON.parse(
-          JSON.stringify(recentListensPropsOneListen)
-        ) as RecentListensProps)}
-      />
-    );
-
-    const instance = wrapper.instance();
-
-    const recordingFeedbackMap: RecordingFeedbackMap = {
-      "973e5620-829d-46dd-89a8-760d87076287": 0,
-    };
-    wrapper.setState(JSON.parse(JSON.stringify(recordingFeedbackMap)));
-
-    await instance.updateFeedback("973e5620-829d-46dd-89a8-760d87076287", 1);
-
-    expect(wrapper.state("recordingFeedbackMap")).toMatchObject({
-      "973e5620-829d-46dd-89a8-760d87076287": 1,
-    });
-  });
-});
-
-describe("updateListenList", () => {
-  it("updates the listens state for particular recording", async () => {
-    /* JSON.parse(JSON.stringify(object) is a fast way to deep copy an object,
-     * so that it doesn't get passed as a reference.
-     */
-    const oneListenProps = JSON.parse(
-      JSON.stringify(recentListensPropsOneListen)
-    );
-    const wrapper = shallow<RecentListens>(
-      <RecentListens {...(oneListenProps as RecentListensProps)} />
-    );
-
-    const instance = wrapper.instance();
-    wrapper.setState({ listens: oneListenProps.listens });
-
-    instance.updateListenList(props.listens[0]);
-    expect(wrapper.state("listens")).toMatchObject([]);
-  });
-});
